strands-template/requirements.txt
python-dotenv
strands
xpander-sdk
openai
anthropic
